fix: adjust version in migration IT
