+++ optimize/upgrade/src/test/java/io/camunda/optimize/upgrade/migrate313to86/AbstractUpgrade86IT.java
@@ -5,20 +5,20 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.optimize.upgrade.migrate313to314;
+package io.camunda.optimize.upgrade.migrate313to86;
 
 import io.camunda.optimize.upgrade.AbstractUpgradeIT;
-import io.camunda.optimize.upgrade.migrate313to314.indices.OnboardingStateIndexV2;
-import io.camunda.optimize.upgrade.migrate313to314.indices.SettingsIndexV2;
+import io.camunda.optimize.upgrade.migrate313to86.indices.OnboardingStateIndexV2;
+import io.camunda.optimize.upgrade.migrate313to86.indices.SettingsIndexV2;
 import io.camunda.optimize.upgrade.plan.UpgradePlan;
 import io.camunda.optimize.upgrade.plan.UpgradePlanRegistry;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 
-public class AbstractUpgrade314IT extends AbstractUpgradeIT {
+public class AbstractUpgrade86IT extends AbstractUpgradeIT {
 
   protected static final String FROM_VERSION = "3.13.0";
-  protected static final String TO_VERSION = "3.14.0";
+  protected static final String TO_VERSION = "8.6.0";
 
   protected final OnboardingStateIndexV2 ONBOARDING_STATE_INDEX = new OnboardingStateIndexV2();
   protected final SettingsIndexV2 SETTINGS_INDEX = new SettingsIndexV2();
