+++ optimize/upgrade/src/test/java/io/camunda/optimize/upgrade/migrate313to86/indices/OnboardingStateIndexV2.java
@@ -5,26 +5,19 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.optimize.upgrade.migrate313to314.indices;
+package io.camunda.optimize.upgrade.migrate313to86.indices;
 
-import static io.camunda.optimize.service.db.DatabaseConstants.OPTIMIZE_DATE_FORMAT;
-
-import io.camunda.optimize.dto.optimize.SettingsDto;
-import io.camunda.optimize.service.db.DatabaseConstants;
 import io.camunda.optimize.service.db.schema.DefaultIndexMappingCreator;
 import java.io.IOException;
 import org.elasticsearch.xcontent.XContentBuilder;
 
-public class SettingsIndexV2 extends DefaultIndexMappingCreator<XContentBuilder> {
+public class OnboardingStateIndexV2 extends DefaultIndexMappingCreator<XContentBuilder> {
 
   public static final int VERSION = 2;
 
-  public static final String SHARING_ENABLED = SettingsDto.Fields.sharingEnabled.name();
-  public static final String LAST_MODIFIED = SettingsDto.Fields.lastModified.name();
-
   @Override
   public String getIndexName() {
-    return DatabaseConstants.SETTINGS_INDEX_NAME;
+    return "onboarding-state";
   }
 
   @Override
@@ -36,18 +29,17 @@ public class SettingsIndexV2 extends DefaultIndexMappingCreator<XContentBuilder>
   public XContentBuilder addProperties(final XContentBuilder xContentBuilder) throws IOException {
     // @formatter:off
     return xContentBuilder
-        .startObject("metadataTelemetryEnabled")
-        .field("type", "boolean")
-        .endObject()
-        .startObject(SHARING_ENABLED)
-        .field("type", "boolean")
-        .endObject()
-        .startObject(LAST_MODIFIED)
-        .field("type", "date")
-        .field("format", OPTIMIZE_DATE_FORMAT)
-        .endObject()
-        .startObject("lastModifier")
+        .startObject("id")
         .field("type", "keyword")
+        .endObject()
+        .startObject("key")
+        .field("type", "keyword")
+        .endObject()
+        .startObject("userId")
+        .field("type", "keyword")
+        .endObject()
+        .startObject("seen")
+        .field("type", "boolean")
         .endObject();
     // @formatter:on
   }
